refactor(tui): move plan history renderer
Extract the plan-update transcript renderer from history.rs into history/plan.rs while preserving the existing PlanUpdateCell re-export and exact rendering behavior.

Refs #3308.
